VisualizationService/requirements.txt
Flask>=1.0.0
azure-cosmos>=1.0.0
cassandra-driver>=1.0.0
pyyaml>=3.13
pillow>=5.4
